add IndexService.cs and BlobService.cs functionality, fix AddCommand.cs

+++ YetAnotherVersionControlSystem/Commands/AddCommand.cs
@@ -1,4 +1,5 @@
-﻿using YetAnotherVersionControlSystem.Contracts;
+﻿using System.Collections;
+using YetAnotherVersionControlSystem.Contracts;
 using YetAnotherVersionControlSystem.Services;
 
 namespace YetAnotherVersionControlSystem.Commands;
@@ -20,15 +21,36 @@ public class AddCommand : ICommand
     }
 
     public string Description { get; set; } = "";
-    
+
+    //TODO попробуй сделать ParallelExecute
+    public void ParallelExecute(params string[] parameters)
+    {
+        var workingDirectory = Environment.CurrentDirectory;
+        var itemFullName = workingDirectory + '/' + parameters[0];
+        if (Directory.Exists(itemFullName))
+        {
+            var paths = new Queue<string>(Directory.GetFileSystemEntries(itemFullName));
+            while (paths.Count != 0)
+            {
+                var currPath = paths.Dequeue();
+            }
+        }
+        else if (File.Exists(itemFullName))
+        {
+            StageFile(itemFullName);
+        }
+        else
+        {
+            throw new Exception("Unknown File Entry");
+        }
+       
+    }
     
     public void Execute(params string[] parameters)
     {
         var workingDirectory = Environment.CurrentDirectory;
-        var vcsRootDirectory = _fileSystemService.GetVcsRootDirectory();
-        var indexFile = vcsRootDirectory.IndexPath;
         var itemFullName = workingDirectory + '/' + parameters[0];
-        
+
         if (File.Exists(itemFullName))
         {
             StageFile(itemFullName);
@@ -41,9 +63,11 @@ public class AddCommand : ICommand
 
     }
 
-    private void StageDirectory(string itemFullName)
+   
+    private void StageDirectory(string directoryPath)
     {
-        var childs = Directory.GetFileSystemEntries(itemFullName);
+        // TODO Add parallelism
+        var childs = Directory.GetFileSystemEntries(directoryPath);
         foreach (var child in childs)
         {
             if (Directory.Exists(child))
@@ -54,42 +78,44 @@ public class AddCommand : ICommand
             {
                 StageFile(child);
             }
-            else throw new Exception("directory include not only files or directory");
+            else throw new Exception("directory include not only files or directories");
         }
         
     }
 
-    private void StageFile(string itemFullName)
+    private void StageFile(string filePath)
     {
-        byte[] bytesArray = File.ReadAllBytes(itemFullName);
+        var bytesArray = File.ReadAllBytes(filePath);
         var hash = _hashService.GetSha1(bytesArray);
-        var indexHash = _indexService.GetHashByPath(itemFullName);
+        var indexHash = _indexService.GetHashByPath(filePath);
             
+        // If file already exist in index
         if (indexHash != null)
         {
+            // if file modified after previous add 
             if (hash != indexHash)
             {
-                //TODO REPLACE
-                _blobService.DeleteBlob(indexHash);
+                // can delete blob if no another files in index have reference to it
+                if (_indexService.GetPathByHash(indexHash).Count == 1)
+                {
+                    _blobService.DeleteBlob(indexHash);
+                }
                 _blobService.CreateBlob(hash, bytesArray);
-                _indexService.DeleteFromIndex(indexHash);
-                _indexService.WriteToIndex(hash, itemFullName);
+                _indexService.DeleteFromIndexByPath(filePath);
+                _indexService.WriteToIndex(hash, filePath);
             }
             else
             {
-                var indexPath = _indexService.GetPathByHash(hash);
-                if (itemFullName != indexPath)
-                {
-                    _indexService.DeleteFromIndex(hash);
-                    _indexService.WriteToIndex(hash, itemFullName);
-                }
-                else throw new Exception("Item is already staged");
+                throw new Exception("Item is already staged");
             }
         }
         else
         {
-            _blobService.CreateBlob(hash,bytesArray);
-            _indexService.WriteToIndex(hash, itemFullName);
+            if (!_blobService.IsBlobExist(hash))
+            {
+                _blobService.CreateBlob(hash, bytesArray);
+            }
+            _indexService.WriteToIndex(hash, filePath);
         }
     }
 }
